reorganize core normalization tests
