+++ packages/slate/test/normalization/editor/remove-text.js
@@ -1,10 +1,10 @@
 /** @jsx jsx */
 
-import { jsx } from '..'
+import { jsx } from '../..'
 
 export const input = (
   <editor>
-    <inline>one</inline>
+    <text>one</text>
     <block>two</block>
   </editor>
 )
