+++ packages/slate/test/normalization/text/merge-adjacent-match.js
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 
-import { jsx } from '..'
+import { jsx } from '../..'
 
 export const schema = {}
