+++ packages/slate/test/normalization/void/inline-insert-text.js
@@ -1,17 +1,12 @@
 /** @jsx jsx */
 
-import { jsx } from '..'
-
-export const schema = {}
+import { jsx } from '../..'
 
 export const input = (
   <editor>
     <block>
       <text />
-      <inline>
-        <block>one</block>
-        <text>two</text>
-      </inline>
+      <inline void />
       <text />
     </block>
   </editor>
@@ -21,8 +16,8 @@ export const output = (
   <editor>
     <block>
       <text />
-      <inline>
-        <text>two</text>
+      <inline void>
+        <text />
       </inline>
       <text />
     </block>
